Remove actix-web dependencies and update Redis and Postgres versions

+++ engine/src/domain/models.rs
@@ -3,7 +3,7 @@ use uuid::Uuid;
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct WagerRequest {
-    pub amount: f64,
+    pub amount: u64,
     pub site_id: i32,
     pub user_id: i32,
     pub game_id: i32,
@@ -16,7 +16,7 @@ pub struct WagerResponse {
     #[serde(with = "uuid::serde::compact")]
     pub wager_id: Uuid,
     pub status: String,
-    pub amount: f64,
+    pub amount: u64,
 
     pub receipt_id: Option<String>,
 }
