+++ storage/src/domain/models.rs
@@ -2,10 +2,10 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct WagerRequest {
-    pub amount: f64,
-    pub site_id: i32,
-    pub user_id: i32,
-    pub game_id: i32,
+    pub amount: u64,
+    pub site_id: u32,
+    pub user_id: u32,
+    pub game_id: u32,
 
     pub cheat_code: Option<String>,
 }
@@ -14,7 +14,7 @@ pub struct WagerRequest {
 pub struct WagerResponse {
     pub wager_id: String,
     pub status: String,
-    pub amount: f64,
+    pub amount: u64,
 
     pub receipt_id: Option<String>,
 }
